Split the command files
