+++ cmd/root.go
@@ -54,6 +54,13 @@ func Execute() {
 	}
 }
 
+func handleError(err error) {
+	if err != nil {
+		fmt.Println(err)
+		os.Exit(1)
+	}
+}
+
 func init() {
 	cobra.OnInitialize(initConfig)
